try pass hexlet-check

+++ app/controllers/likes_controller.rb
@@ -19,10 +19,8 @@ class LikesController < ApplicationController
   end
 
   def destroy
-    if current_user
-      like = PostLike.find(params[:id])
-      like.destroy
-    end
+    like = PostLike.find(params[:id])
+    like.destroy if current_user&.likes&.any? { |users_like| users_like == like }
     # не работает с turbo. без turbo не пробовал
     # redirect_to "#{request.referer}#post_#{post_id}"
     redirect_to request.url
